renamed home screen to welcome screen
